+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fix legends typings [#335](https://github.com/CartoDB/carto-react/pull/335)
 - Fix legends export [#334](https://github.com/CartoDB/carto-react/pull/334)
 - Fix duplicated logic for category selection in PieWidgetUI [#332](https://github.com/CartoDB/carto-react/pull/332)
+- Add customLegendTypes prop to customise covered legend types [#325](https://github.com/CartoDB/carto-react/pull/325)
+- Add error label if legend type is unknown [#325](https://github.com/CartoDB/carto-react/pull/325)
 
 ## 1.2
 
@@ -15,7 +17,6 @@
 - Fix error in getPalette, causing wrong colors when using CARTOColors [#328](https://github.com/CartoDB/carto-react/pull/328)
 - Fix errors in LegendRamp with wrong labels [#328](https://github.com/CartoDB/carto-react/pull/328)
 - Dynamic tiling integration [#327](https://github.com/CartoDB/carto-react/pull/327)
-- Add customLegendTypes prop to customise covered legend types [#325](https://github.com/CartoDB/carto-react/pull/325)
 - Fix Switch input width [#323](https://github.com/CartoDB/carto-react/pull/323)
 - Add optional slider to modify layer opacity in Legend widget [#324](https://github.com/CartoDB/carto-react/pull/324)