+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Not released
 
 - Set expanded/collapsed value of each legend row [#337](https://github.com/CartoDB/carto-react/pull/337)
+- Use layerConfig in useCartoLayerProps [#336](https://github.com/CartoDB/carto-react/pull/336)
 - Fix legends typings [#335](https://github.com/CartoDB/carto-react/pull/335)
 - Fix legends export [#334](https://github.com/CartoDB/carto-react/pull/334)
 - Fix duplicated logic for category selection in PieWidgetUI [#332](https://github.com/CartoDB/carto-react/pull/332)